fast-api-app/requirements.txt
uvicorn
pydantic 
typing-extensions
python-dateutil
pyyaml
python-json-logger
fastapi[standard]
semantic-router[openai]
ollama
kagglehub[pandas-datasets]